Potential fix for issue #218
The issue https://github.com/websocket-rails/websocket-rails/issues/218 is about tests not triggering any actions. This is caused because the execute method in dispatcher.rb processes actions asynchronously with EM.next_tick. This causes all tests to fail as the action is never executed.

We deal with this by checking if this is the test environment with Rails.env.test? and executing the actions immediately in test environment.